OP-2860 - extracted get_fps function to lib
