+++ openpype/lib/vendor_bin_utils.py
@@ -130,3 +130,23 @@ def is_oiio_supported():
         ))
         return False
     return True
+
+
+def get_fps(str_value):
+    """Returns (str) value of fps from ffprobe frame format (120/1)"""
+    if str_value == "0/0":
+        print("WARNING: Source has \"r_frame_rate\" value set to \"0/0\".")
+        return "Unknown"
+
+    items = str_value.split("/")
+    if len(items) == 1:
+        fps = float(items[0])
+
+    elif len(items) == 2:
+        fps = float(items[0]) / float(items[1])
+
+    # Check if fps is integer or float number
+    if int(fps) == fps:
+        fps = int(fps)
+
+    return str(fps)
